Test for #700
